Blinding factor returned when create transp note
The blinding factor is a required return of the obfuscated note
creation.

To maintain API consistency, the creation of the transparent note should
return the blinding factor as well.

It happens to be constant zero. But, to keep this logic decoupled from
the consumers of this library, and if ever this constant changes, its
logic is now granted to be defined only inside of phoenix-core