Blinding factor as parameter of Note::obfuscated
The blinding factor is required to prove the knowledge of the value
commitment of the note.

If the blinding factor is not provided by the note creator, the latter
will not be able to fetch this information, and consequentially won't be
able to create a zk proof of knownledge of the value commitment without
the view key of the owner of the note.

This is not desirable since a note should be created only with the
public key of the receiver.