Closes #65 - Fee - Out of gas exceptions
Prior to this commit, no gas limit or gas consumed validations were
performed. This can lead to several bugs in phoenix implementations, and
there is no reason to not perform the checks in this library since it
contains all the required data for every validation.

Also, a consumed fee note will need to be created from a given stealth
address to a block generator. Prior to this commit, the API will not
allow note creation with a specific stealth address without the
knowledge of `r`.