crossover: use all attributes as hash inputs
The hash representation of the crossover should reflect all the
attributes.

Even if the crossover is composed only by `value` and `blinder`, and
both are present in the value commitment, its hash representation should
also guarantee the consistency of the whole object. This way, we
increase the overall security of the system.

Resolves #69