Update canonical from v0.5 to v0.6 (#73)
Include the new features and simplification of Canonical API.

With the update, its viable to use standard representations for the
error type.

The latest dusk libraries deprecated the usage of `rand::thread_rng` in
favor os `rand_core::OsRng`. This commit also update the test cases
accordingly and move the tests of the message to the proper test
structure.

Resolves #72