Add `Note::transparent_stealth` function
The function allows the generation of a note by using just a stealth
address, a value, and a nonce, in contrast with `Note::transparent`
which requires a public spend key.
This is required for environments where knowledge of the public spend
key is undesirable, but a stealth address is acceptable.