Update `dusk-poseidon` dependency to `0.28`
This requires the removal of the `canon` feature, since `dusk-poseidon`
stopped providing it, and the removal of the accompanying error. Since
decryption *can* error however - signalled by an `Option` from
`dusk-poseidon`, a new error type called `Decryption` is also
introduced.

I also took the opportunity to rename the `BadLenght` error to `BadLength`.

Resolves #114